Returns -1 when legacy getchar function fails.
Ref: https://github.com/riscv-software-src/opensbi/pull/76/files#diff-a3ecbce4af91d3e4271388be8f36b1bbbb401baf37b76f1caadaa119d1110127R51-R56 and SBI specification section 4.3

+++ src/ecall/legacy.rs
@@ -11,10 +11,11 @@ pub fn console_putchar(param0: usize) -> SbiRet {
     SbiRet::ok(0) // the return value 0 is ignored in legacy
 }
 
+// Returns a character, or -1 when failed.
 #[inline]
 pub fn console_getchar() -> SbiRet {
     let ch = legacy_stdio_getchar();
-    SbiRet::legacy_ok(ch as usize)
+    SbiRet::legacy_ok(ch)
 }
 
 #[inline]
